[ADD]-l-modifier

+++ lib/my/my_memset.c
@@ -14,7 +14,7 @@ void *my_memset(void *ptr, int j, size_t k)
 {
     char *temp = malloc(sizeof(char) * (my_strlen(ptr) + 1));
 
-    for (int i = 0; i < k; i++)
+    for (size_t i = 0; i < k; i++)
         temp[i] = j;
     return (temp);
 }
